Rewrite cis1 examples to cis2 draft version 1
