+++ .github/workflows/linter.yml
@@ -20,18 +20,14 @@ jobs:
           - concordium-std-derive/Cargo.toml
           - concordium-std/Cargo.toml
           - examples/auction/Cargo.toml
-          - examples/cis1-multi/Cargo.toml
-          - examples/cis1-nft/Cargo.toml
-          - examples/cis1-single-nft/Cargo.toml
-          - examples/cis1-wccd/Cargo.toml
-          - examples/erc721/Cargo.toml
+          - examples/cis2-multi/Cargo.toml
+          - examples/cis2-nft/Cargo.toml
+          - examples/cis2-wccd/Cargo.toml
           - examples/fib/Cargo.toml
           - examples/icecream/Cargo.toml
-          - examples/listing-cis1/Cargo.toml
           - examples/memo/Cargo.toml
           - examples/piggy-bank/part1/Cargo.toml
           - examples/piggy-bank/part2/Cargo.toml
-          - examples/proxy/Cargo.toml
           - examples/proxy/Cargo.toml
           - examples/recorder/Cargo.toml
           - examples/signature-verifier/Cargo.toml
@@ -215,10 +211,9 @@ jobs:
 
         crates:
           - examples/two-step-transfer/Cargo.toml
-          - examples/erc721/Cargo.toml
-          - examples/cis1-wccd/Cargo.toml
-          - examples/cis1-nft/Cargo.toml
-          - examples/cis1-multi/Cargo.toml
+          - examples/cis2-wccd/Cargo.toml
+          - examples/cis2-nft/Cargo.toml
+          - examples/cis2-multi/Cargo.toml
 
         features:
           -
@@ -291,18 +286,14 @@ jobs:
 
         crates:
           - examples/auction/Cargo.toml
-          - examples/cis1-multi/Cargo.toml
-          - examples/cis1-nft/Cargo.toml
-          - examples/cis1-single-nft/Cargo.toml
-          - examples/cis1-wccd/Cargo.toml
-          - examples/erc721/Cargo.toml
+          - examples/cis2-multi/Cargo.toml
+          - examples/cis2-nft/Cargo.toml
+          - examples/cis2-wccd/Cargo.toml
           - examples/fib/Cargo.toml
           - examples/icecream/Cargo.toml
-          - examples/listing-cis1/Cargo.toml
           - examples/memo/Cargo.toml
           - examples/piggy-bank/part1/Cargo.toml
           - examples/piggy-bank/part2/Cargo.toml
-          - examples/proxy/Cargo.toml
           - examples/proxy/Cargo.toml
           - examples/recorder/Cargo.toml
           - examples/signature-verifier/Cargo.toml
@@ -376,18 +367,14 @@ jobs:
 
         crates:
           - examples/auction/Cargo.toml
-          - examples/cis1-multi/Cargo.toml
-          - examples/cis1-nft/Cargo.toml
-          - examples/cis1-single-nft/Cargo.toml
-          - examples/cis1-wccd/Cargo.toml
-          - examples/erc721/Cargo.toml
+          - examples/cis2-multi/Cargo.toml
+          - examples/cis2-nft/Cargo.toml
+          - examples/cis2-wccd/Cargo.toml
           - examples/fib/Cargo.toml
           - examples/icecream/Cargo.toml
-          - examples/listing-cis1/Cargo.toml
           - examples/memo/Cargo.toml
           - examples/piggy-bank/part1/Cargo.toml
           - examples/piggy-bank/part2/Cargo.toml
-          - examples/proxy/Cargo.toml
           - examples/proxy/Cargo.toml
           - examples/recorder/Cargo.toml
           - examples/signature-verifier/Cargo.toml
