--- concordium-cis1/CHANGELOG.md
@@ -1,5 +1,0 @@
-# Changelog
-
-## Unreleased changes
-
-   - Fix a bug in schema type for transfer parameters.
