+++ concordium-cis2/CHANGELOG.md
@@ -1,0 +1,3 @@
+# Changelog
+
+
