+++ examples/cis2-wccd/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "cis1-nft"
+name = "cis2_wccd"
 version = "0.1.0"
 authors = ["Concordium <developers@concordium.com>"]
 edition = "2018"
@@ -7,15 +7,15 @@ license = "MPL-2.0"
 
 [features]
 default = ["std"]
-std = ["concordium-std/std", "concordium-cis1/std"]
+std = ["concordium-std/std", "concordium-cis2/std"]
 
 [dependencies]
 concordium-std = {path = "../../concordium-std", default-features = false}
-concordium-cis1 = {path = "../../concordium-cis1", default-features = false}
+concordium-cis2 = {path = "../../concordium-cis2", default-features = false}
 
 [lib]
 crate-type=["cdylib", "rlib"]
 
 [profile.release]
-codegen-units = 1
 opt-level = "s"
+codegen-units = 1
